0x101 의 8칸을 전부 float64 로 맞춘다 — 마지막 칸만 u64 였다
dSPACE 는 이 프레임을 **float64 8칸**으로 읽는다(팀장 확인, 2026-08-29). main 은
이미 64B·little-endian·FD/BRS 로 그 배열을 보내고 있었지만 마지막 `update` 만
uint64_t 로 선언돼 있었다. 값이 0 인 동안은 두 자료형의 바이트가 같아 드러나지
않지만, 거기에 정수를 넣는 순간 상대가 double 로 읽어 엉뚱한 값이 된다.
와이어를 바꾸는 변경이 아니라 그 사고를 자료형으로 막는 변경이다.

현장 검증 (Kvaser Leaf v3, can0, 1M/2M):
  · 0x101 에 f64 x8 = 1.0~8.0 을 10ms 로 500프레임 송신 → 전부 ACK, 오류 0
  · dSPACE 회신 0x200 도 같은 형식(64B·FD·BRS·100Hz)으로 수신
  · protocol_selftest.py ALL PASS

+++ src/bridge_dspace/src/can_protocol.hpp
@@ -97,10 +97,15 @@ struct MpcTargetFdPayload  // 0x101 MPC_TARGET_FD — 64 B
   double curvature;  // [1/m]
   // dx/dy/dyaw/update: PR #52 에서 의미 미정. **0 으로 채운다** (팀장 결정 2026-08-28).
   // 의미가 정해지면 여기와 PROTOCOL.md 를 함께 갱신할 것.
+  //
+  // 2026-08-29: `update` 를 u64 → **double** 로 맞췄다. 이 프레임은 dSPACE 쪽에서
+  // **float64 8칸**으로 읽는 것이 계약이다(팀장 확인). 값이 0 인 동안은 두 자료형의
+  // 바이트가 같아 지금 당장은 와이어가 안 바뀌지만, `update` 에 정수를 넣는 순간
+  // 상대가 double 로 읽어 엉뚱한 값이 된다 — 그 사고를 자료형으로 막는다.
   double dx;
   double dy;
   double dyaw;
-  uint64_t update;
+  double update;
 };
 
 struct VehFeedbackFdPayload  // 0x200 VEH_FEEDBACK_FD — 64 B
